+++ PhpSerializerNET.Test/TestDeserializeErrors.cs
@@ -12,25 +12,25 @@ namespace PhpSerializerNET.Test {
 		[TestMethod]
 		public void ThrowsOnMalformedNull() {
 			var ex = Assert.ThrowsException<DeserializationException>(() => PhpSerialization.Deserialize("N"));
-			Assert.AreEqual("Malformed null at position 0", ex.Message);
+			Assert.AreEqual("Malformed null at position 0: Expected ';'", ex.Message);
 
 			ex = Assert.ThrowsException<DeserializationException>(() => PhpSerialization.Deserialize("N?"));
-			Assert.AreEqual("Malformed null at position 0", ex.Message);
+			Assert.AreEqual("Malformed null at position 0: Expected ';'", ex.Message);
 		}
 
 		[TestMethod]
 		public void ThrowsOnMalformedBool() {
 			var ex = Assert.ThrowsException<DeserializationException>(() => PhpSerialization.Deserialize("b"));
 			Assert.IsNotNull(ex);
-			Assert.AreEqual("Malformed boolean at position 0", ex.Message);
+			Assert.AreEqual("Malformed boolean at position 0: Unexpected end of input sequence.", ex.Message);
 
 			ex = Assert.ThrowsException<DeserializationException>(() => PhpSerialization.Deserialize("b?"));
 			Assert.IsNotNull(ex);
-			Assert.AreEqual("Malformed boolean at position 0", ex.Message);
+			Assert.AreEqual("Malformed boolean at position 0: Unexpected end of input sequence.", ex.Message);
 
 			ex = Assert.ThrowsException<DeserializationException>(() => PhpSerialization.Deserialize("b:1"));
 			Assert.IsNotNull(ex);
-			Assert.AreEqual("Malformed boolean at position 0", ex.Message);
+			Assert.AreEqual("Malformed boolean at position 0: Unexpected end of input sequence.", ex.Message);
 		}
 
 		[TestMethod]
@@ -48,19 +48,19 @@ namespace PhpSerializerNET.Test {
 			Assert.AreEqual("Malformed string at position 0", ex.Message);
 
 			ex = Assert.ThrowsException<DeserializationException>(() => PhpSerialization.Deserialize("s:1:\"a "));
-			Assert.AreEqual("Malformed string at position 0", ex.Message);
+			Assert.AreEqual("String at position 4 has an incorrect length of 1: Expected double quote at position 6, found ' ' instead.", ex.Message);
 
 			ex = Assert.ThrowsException<DeserializationException>(() => PhpSerialization.Deserialize("s:1:\"a"));
 			Assert.AreEqual("Malformed string at position 0", ex.Message);
 
 			ex = Assert.ThrowsException<DeserializationException>(() => PhpSerialization.Deserialize("s:30:\"\";"));
-			Assert.AreEqual("Illegal length of 30. The string at position 0 points to out of bounds index 36.", ex.Message);
+			Assert.AreEqual("Illegal length of 30. The string at position 4 points to out of bounds index 36.", ex.Message);
 
 			ex = Assert.ThrowsException<DeserializationException>(() => PhpSerialization.Deserialize("s:1:\"\";"));
-			Assert.AreEqual("String at position 0 has an incorrect length of 1: Expected double quote at position 6, found ';' instead.", ex.Message);
+			Assert.AreEqual("String at position 4 has an incorrect length of 1: Expected double quote at position 6, found ';' instead.", ex.Message);
 
 			ex = Assert.ThrowsException<DeserializationException>(() => PhpSerialization.Deserialize("s:a:\"\";"));
-			Assert.AreEqual("Malformed string at position 0", ex.Message);
+			Assert.AreEqual("String at position 3 has illegal, missing or malformed length.", ex.Message);
 		}
 
 		[TestMethod]