+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Future:
+- Further performance improvements.
+- Improved exception messages on malformed input when deserializing.
+
 # 0.7.4:
 - Improved deserialization performance.
 - Fixed invalid output when using PhpSerializiationOptions.NumericEnums = false